refactor : 카테고리 선택 버튼, 드롭 다운 메뉴를 감싸는 CategorySelector 구현
- 카테고리 선택 버튼 컴포넌트명 변경 (SearchCategoryButton -> CategorySelectorButton)
