+++ src/components/Searchbar/components/CategorySelector/components/DropdownMenu/index.module.scss
@@ -1,4 +1,4 @@
-@use '../../_variable' as vars;
+@use '@/components/Searchbar/_variable' as vars;
 
 .dropdownMenu {
   position: absolute;
