+++ src/components/Searchbar/components/CategorySelector/hooks/useOpenDropdownMenu.ts
