+++ src/components/Searchbar/index.tsx
@@ -1,17 +1,15 @@
 'use client';
 
 import Image from 'next/image';
-import React, { useActionState, useRef, useState } from 'react';
+import React, { useActionState, useRef } from 'react';
 
 import SearchIcon from '@/images/searchIcon.svg';
 
 import { Toast } from '../common';
 
 import searchAction from './action/searchAction';
-import DropdownMenu from './components/DropdownMenu';
-import SearchCategoryButton from './components/SearchCategoryButton';
+import CategorySelector from './components/CategorySelector';
 import useElementId from './hooks/useElementId';
-import useOpenDropdownMenu from './hooks/useOpenDropdownMenu';
 import useToast from './hooks/useToast';
 import styles from './index.module.scss';
 
@@ -20,26 +18,11 @@ interface Props {
 }
 
 const Searchbar = ({ categoryInfo }: Props) => {
-  const [selectedCategory, setSelectedCategory] = useState('title');
   const [state, formAction, isPending] = useActionState(searchAction, null);
   const { isOpenToast, handleCloseToast } = useToast({ state });
 
   const elementId = useElementId();
   const searchInputRef = useRef<HTMLInputElement>(null);
-
-  const {
-    dropdownMenuButtonRef,
-    dropdownMenuRef,
-    isOpenDropdownMenu,
-    handleClickDropdownOpenButton,
-    closeDropdownMenu,
-  } = useOpenDropdownMenu();
-
-  const changeCategory = (item: string) => {
-    if (item === selectedCategory) return;
-    setSelectedCategory(item);
-    closeDropdownMenu();
-  };
 
   const handleFormAction = (payload: FormData) => {
     if (!searchInputRef.current || searchInputRef.current.value.trim() === '') return;
@@ -48,14 +31,7 @@ const Searchbar = ({ categoryInfo }: Props) => {
 
   return (
     <form className={styles.searchbar} action={handleFormAction}>
-      <SearchCategoryButton
-        dropdownMenuButtonRef={dropdownMenuButtonRef}
-        isOpenDropdownMenu={isOpenDropdownMenu}
-        elementId={elementId}
-        handleClickDropdownOpenButton={handleClickDropdownOpenButton}
-        categoryInfo={categoryInfo}
-        selectedCategory={selectedCategory}
-      />
+      <CategorySelector elementId={elementId} categoryInfo={categoryInfo} />
       <label className="sr-only" htmlFor={elementId.searchInput}>
         검색어
       </label>
@@ -70,15 +46,7 @@ const Searchbar = ({ categoryInfo }: Props) => {
       <button type="submit" className={styles.searchButton} disabled={isPending}>
         <Image src={SearchIcon} alt="" width={17} height={18} />
       </button>
-      {isOpenDropdownMenu && (
-        <DropdownMenu
-          id={elementId.searchCategoryList}
-          dropdownMenuRef={dropdownMenuRef}
-          selectedCategory={selectedCategory}
-          categoryInfo={categoryInfo}
-          changeCategory={changeCategory}
-        />
-      )}
+
       {isOpenToast && (
         <Toast handleCloseToast={handleCloseToast}>
           <p>{state?.error}</p>
